Fix parsing entire output file when -O is combined with recursion
When recursion is enabled, the downloaded file will be parsed for
additional URLs. When recursive downloads are concatenated in a single
file due to the -O option, the entire file is parsed again. Fix this by
seeking to the point in the output file before the last downloaded file
before parsing.

Signed-off-by: John Brooks <john@fastquake.com>