Fix re-parsing entire output file when -O is combined with recursion
When recursion is enabled, the downloaded file will be parsed for
additional URLs. When recursive downloads are concatenated in a single
file due to the -O option, the entire file is parsed again. This can
severely impact performance; I observed delays of tens of seconds when the
output file reached a size of about 20 MB.

Fix this by seeking to the point in the output file where the last
downloaded file begins prior to parsing out URLs.

Signed-off-by: John Brooks <john@fastquake.com>